requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
pydantic>=2.8.0
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.7.1
pytest>=8.3.2
